Fix clang builds in wscript
Clang builds were defaulting to g++ builds
as the 'variant' attribute of the 'clang_debug'
subclass of BuildContext was not what the clang configuration
method had 'setenv' to. As a result clang buildsd defaulted
to g++, instead of compile failure.
The linker flag '-glidb' was wrong, but not caught in clang
builds. I fixed it to '-glldb'. It is some kind of a debugger
link.